requirements.txt
pandas==1.5.2
salesforce_api==0.1.45
salesforce_bulk==2.2.0
tqdm==4.64.1
unicodecsv==0.14.1
python-dotenv
pyzipper